requirements.txt
transformers>=4.35.0
torch>=2.0.0
parrot-paraphraser>=0.0.1
gramformer>=1.0
nlpaug>=1.1.11
spacy>=3.6.0
textacy>=0.13.0
scipy>=1.11.0
scikit-learn>=1.3.0
nltk>=3.8.1
wordfreq>=3.1.0
sentence-transformers>=2.2.2
deep-translator>=1.11.4
textblob>=0.17.1
vaderSentiment>=3.3.2
language-tool-python>=2.7.1
proselint>=0.14.0
python-docx>=1.0.1
reportlab>=4.0.4
habanero>=1.2.3
citeproc-py>=0.6.0
pybtex>=0.24.0
sumy>=0.11.0
unidecode>=1.3.7
pandas>=2.0.3
requests>=2.31.0
beautifulsoup4>=4.12.2
pyyaml>=6.0.1
fonttools>=4.42.1
keybert>=0.8.3
PyMultiDictionary>=1.2.0
latexify-py>=0.4.0
PyICU>=2.10.2
handcalcs>=1.7.0
tabulate>=0.9.0
